sdcard_avr.c: another critical bug fix of SDHC/SDXC ccs=1 mode
The last fix of 56c1d6026ceff9edc2bd139c64269e06a7165824 was not
sufficient. Actually, without the last fix, ccs=1 mode worked
correctly by chance, but ccs=0 mode was broken.

This was because the condition check in sdcard_store was inverted.
As a result, if ccs=1, cached block address was used as is, thus
it worked. But if ccs=0, cached raw address was shifted by 9 bit,
then it broke the first block.

After my last fix, both mode have not worked, but with this second
fix, everything should work correctly.